+++ CHANGES.md
@@ -23,6 +23,7 @@
 * Add authentication for CQL and JMX - Issue #129
 * Add TLS support for CQL and JMX - Issue #129
 * Expose Springboot configuration - Issue #149
+* Per table configurations - Issue #119
 
 ### Merged from 1.2